+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.3.0
+
+- Added `TExtendable` to `EStates`.
+- Added method `IInstanceService::updateInstance(IInstance $instance, array $data, array $options): bool`.
+
 # 1.2.1
 
 - Rm unused code.